Move some tests out of test_multidevice.py
